+++ app/src/assets/main.css
@@ -1,35 +1,39 @@
 @import './base.css';
 
+nav {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: 8px;
+  background: var(--vt-c-black-mute);
+  padding: 8px 12px;
+  border-radius: 12px;
+
+  :first-child {
+    margin-inline-end: auto;
+  }
+
+  a {
+    padding: 8px;
+    background: var(--vt-c-black-soft);
+    border-radius: 4px;
+  }
+}
+
 #app {
   max-width: 1280px;
   margin: 0 auto;
-  padding: 2rem;
+  padding: 16px;
   font-weight: normal;
 }
 
-a,
-.green {
+a {
   text-decoration: none;
   color: hsla(160, 100%, 37%, 1);
   transition: 0.4s;
   padding: 3px;
 }
 
-@media (hover: hover) {
-  a:hover {
-    background-color: hsla(160, 100%, 37%, 0.2);
-  }
-}
-
-@media (min-width: 1024px) {
-  body {
-    display: flex;
-    place-items: center;
-  }
-
-  #app {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    padding: 0 2rem;
-  }
+main {
+  padding: 16px;
 }
