+++ app/src/views/AboutView.vue
@@ -1,15 +1,17 @@
 <template>
-  <div class="about">
-    <h1>This is an about page</h1>
-  </div>
+  <h1>About BinTraq</h1>
+  <p><i>A simple bin record keeping system for orchards.</i></p>
+  <p>
+    Developed by <a href="https://kwila.cloud">Kwila Cloud</a> for
+    <a href="https://traesureappleranch.com">Traesure Apple Ranch.</a>
+  </p>
+  <p>
+    For more information, see our <a href="https://github.com/kwila-cloud/bintraq">GitHub page</a>.
+  </p>
 </template>
 
 <style>
-@media (min-width: 1024px) {
-  .about {
-    min-height: 100vh;
-    display: flex;
-    align-items: center;
-  }
+p {
+  padding: 8px 0;
 }
 </style>
