+++ app/src/views/HomeView.vue
@@ -1,9 +1,29 @@
 <script setup lang="ts">
-import TheWelcome from '../components/TheWelcome.vue'
+import { supabase } from '@/lib/supabaseClient'
+import { ref, onMounted } from 'vue'
+
+type Bin = {
+  uuid: string
+  id: string
+  picker: string
+}
+
+const bins = ref<Bin[]>([])
+
+async function getBins() {
+  const { data } = await supabase.from('bin').select().order('date')
+  bins.value = data as Bin[]
+}
+
+onMounted(() => {
+  getBins()
+})
 </script>
 
 <template>
-  <main>
-    <TheWelcome />
-  </main>
+  <h1>Bins</h1>
+  <ul>
+    <li v-for="bin in bins" :key="bin.uuid">{{ bin.id }} (picked by {{ bin.picker }})</li>
+  </ul>
+  <p style="padding-top: 16px">Authentication guard coming soon...</p>
 </template>
